+++ computer-core/src/main/java/com/baidu/hugegraph/computer/core/input/InputSourceFactory.java
@@ -24,8 +24,8 @@ import com.baidu.hugegraph.computer.core.config.ComputerOptions;
 import com.baidu.hugegraph.computer.core.config.Config;
 import com.baidu.hugegraph.computer.core.input.hg.HugeGraphFetcher;
 import com.baidu.hugegraph.computer.core.input.hg.HugeInputSplitFetcher;
-import com.baidu.hugegraph.computer.core.input.loader.FileGraphFetcher;
-import com.baidu.hugegraph.computer.core.input.loader.HdfsInputSplitFetcher;
+import com.baidu.hugegraph.computer.core.input.loader.LoaderFileInputSplitFetcher;
+import com.baidu.hugegraph.computer.core.input.loader.LoaderGraphFetcher;
 import com.baidu.hugegraph.computer.core.rpc.InputSplitRpcService;
 
 public class InputSourceFactory {
@@ -35,8 +35,8 @@ public class InputSourceFactory {
         switch (type) {
             case "hugegraph":
                 return new HugeInputSplitFetcher(config);
-            case "hdfs":
-                return new HdfsInputSplitFetcher(config);
+            case "loader":
+                return new LoaderFileInputSplitFetcher(config);
             default:
                 throw new ComputerException("Unexpected source type %s", type);
         }
@@ -48,8 +48,8 @@ public class InputSourceFactory {
         switch (type) {
             case "hugegraph":
                 return new HugeGraphFetcher(config, srv);
-            case "hdfs":
-                return new FileGraphFetcher(config, srv);
+            case "loader":
+                return new LoaderGraphFetcher(config, srv);
             default:
                 throw new ComputerException("Unexpected source type %s", type);
         }
