+++ computer-core/src/main/java/com/baidu/hugegraph/computer/core/input/loader/LoaderGraphFetcher.java
@@ -26,15 +26,15 @@ import com.baidu.hugegraph.computer.core.input.InputSplit;
 import com.baidu.hugegraph.computer.core.input.VertexFetcher;
 import com.baidu.hugegraph.computer.core.rpc.InputSplitRpcService;
 
-public class FileGraphFetcher implements GraphFetcher {
+public class LoaderGraphFetcher implements GraphFetcher {
 
     private final InputSplitRpcService rpcService;
     private final VertexFetcher vertexFetcher;
     private final EdgeFetcher edgeFetcher;
 
-    public FileGraphFetcher(Config config, InputSplitRpcService rpcService) {
-        this.vertexFetcher = new FileVertxFetcher(config);
-        this.edgeFetcher = new FileEdgeFetcher(config);
+    public LoaderGraphFetcher(Config config, InputSplitRpcService rpcService) {
+        this.vertexFetcher = new LoaderVertxFetcher(config);
+        this.edgeFetcher = new LoaderEdgeFetcher(config);
         this.rpcService = rpcService;
     }
